Added error handler
